Create README.md
I have all the skills required to push the community to greater heights. I believe my experience and expertise will be a very good asset to the company. I'm a team player and a goal getter. I believe in community and the value it stands for so I am more than able to work to the success of whatever field I am stationed.